大二下/软件综合设计/voicebar-backend-develop/requirements.txt
asgiref
authing
certifi
charset-normalizer
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
gql
graphql-core
idna
mysqlclient
promise
pyasn1
PyJWT
python-dateutil
pytz
requests
rsa
Rx
six
sqlparse
tzdata
urllib3
azure-cognitiveservices-speech
Pillow
cos-python-sdk-v5
qqwry-py3